Bluetooth: btmtk: move btusb_mtk_[setup, shutdown] to btmtk.c
Move btusb_mtk_[setup, shutdown] and related function from
btusb.c to btmtk.c which holds vendor specific stuff and
would make btusb.c clean.

(cherry picked from commit 5c5e8c52e3cafaa6f71efd44a42a674271e5b3ad)
